This instrument is driven by motion capture from webcam input
Using cinder, openCV, and OSC, this project detects motion with a
webcam through frame differencing. It searches for a grid area with the
maximum motion and passes the corresponding values to a Processing file
through OSC. The values passed are the sum of the overall amount of
“motion” in the grid (done through pixel addition), the area’s x value,
and the area’s y value. These three values drive a triangle oscillator
in order to create instrument that you “play” through movement.